// src/enums/LocalStorageKeys.ts
enum LocalStorageKeys {
  CREATED_AT = "created_at",
  TOKEN = "token_saude_braspires",
  USER_ID = "user_id",
  EMAIL = "email",
  NOME = "nome",
  FOTO = "foto",
  RUA = "rua",
  NUMERO = "numero",
  CIDADE = "cidade",
  BAIRRO = "bairro",
  PONTO_DE_REFERENCIA = "ponto_de_referencia",
}
export { LocalStorageKeys };